+++ src/entities/mypage-scrap/ui/scrap-list.tsx
@@ -5,7 +5,7 @@ import { useFilterStore } from "@/src/features/filter-announcements/model/use-fi
 import { AnnouncementCard } from "@/src/widgets/announcement-card";
 import { useQuery } from "@tanstack/react-query";
 import { useMemo } from "react";
-import { getScrappedAnnouncements } from "../api/quries";
+import { getScrappedAnnouncements } from "../api/mypage-scrap.quries";
 
 export function ScrapList() {
   const { appliedFilters } = useFilterStore();
