+++ src/entities/user/index.ts
@@ -1,4 +1,4 @@
-export { useUserStore } from "@/src/entities/user/model/user-store";
-export { getProfile } from "@/src/entities/user/api/get-profile";
+export { useUserStore } from "@/src/entities/user/model/use-user-store";
+export { getProfile } from "@/src/entities/user/api/profile.action";
 export type { User } from "@/src/entities/user/model/user.types";
 export { useUser } from "@/src/entities/user/lib/use-user";
